requirements.txt
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.2
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
psycopg2-binary==2.9.3
pylint==2.12.2
black==22.1.0
django-cors-headers==3.11.0
django-filter==21.1
django-environ==0.8.1
django-pgconnection==1.0.2
django-extensions==3.1.5
django-cors-headers==3.11.0
drf-yasg==1.20.0
